requirements.txt
python-decouple
python-telegram-bot
pymupdf
filetype
peewee